+++ routes/clients.js
@@ -1,0 +1,3 @@
+//TODO: GET ACCESS VIA MIDDLEWARE
+//TODO: LOGIC FOR POSTS AND COMMENTS --> POST
+//TODO: LOGIC FOR UPVOTES/LIKES/RATINGS --> POST
